refactor transform
