+++ teamster/core/ops/datagun.py
@@ -1,6 +1,79 @@
+import gzip
+import json
 import pathlib
 
-from dagster import In, RetryPolicy, Tuple, op
+import pandas as pd
+from dagster import Dict, In, List, Out, Output, RetryPolicy, Tuple, op
+
+from teamster.core.utils import NOW, TODAY, CustomJSONEncoder, get_last_schedule_run
+
+
+@op(
+    ins={
+        "data": In(dagster_type=List[Dict]),
+        "file_config": In(dagster_type=Dict),
+        "dest_config": In(dagster_type=Dict),
+    },
+    out={"transformed": Out(dagster_type=Tuple, is_required=False)},
+    required_resource_keys={"file_manager"},
+    tags={"dagster/priority": 3},
+)
+def transform(context, data, file_config, dest_config):
+    mapping_key = context.get_mapping_key()
+    table_name = mapping_key[: mapping_key.rfind("_")]
+
+    now_ts = NOW.timestamp()
+    if file_config:
+        file_stem = file_config["stem"].format(
+            today=TODAY.date().isoformat(),
+            now=now_ts.replace(".", "_"),
+            last_run=get_last_schedule_run(context=context),
+        )
+        file_suffix = file_config["suffix"]
+        file_format = file_config.get("format", {})
+
+        file_encoding = file_format.get("encoding", "utf-8")
+    else:
+        file_stem = now_ts.replace(".", "_")
+        file_suffix = "json.gz"
+        file_format = {}
+        file_encoding = "utf-8"
+
+    dest_type = dest_config["type"]
+
+    if dest_type == "gsheet":
+        context.log.info("Transforming data to DataFrame")
+        df = pd.DataFrame(data=data)
+        df_json = df.to_json(orient="split", date_format="iso", index=False)
+
+        df_dict = json.loads(df_json)
+        df_dict["shape"] = df.shape
+
+        yield Output(value=(dest_config, file_stem, df_dict), output_name="transformed")
+    elif dest_type in ["fs", "sftp"]:
+        context.log.info(f"Transforming data to {file_suffix}")
+
+        if file_suffix == "json":
+            data_bytes = json.dumps(obj=data, cls=CustomJSONEncoder).encode(
+                file_encoding
+            )
+        elif file_suffix == "json.gz":
+            data_bytes = gzip.compress(
+                json.dumps(obj=data, cls=CustomJSONEncoder).encode(file_encoding)
+            )
+        elif file_suffix in ["csv", "txt", "tsv"]:
+            df = pd.DataFrame(data=data)
+            data_bytes = df.to_csv(index=False, **file_format).encode(file_encoding)
+
+        file_handle = context.resources.file_manager.write_data(
+            data=data_bytes,
+            key=f"{(dest_config.get('name') or table_name or 'data')}/{file_stem}",
+            ext=file_suffix,
+        )
+        context.log.info(f"Saved to {file_handle.path_desc}.")
+
+        if dest_type == "sftp":
+            yield Output(value=(dest_config, file_handle), output_name="transformed")
 
 
 @op(
